requirements.txt
# Core Testing
pytest==9.0.2
pytest-xdist==3.8.0
pytest-html==4.2.0
pytest-rerunfailures==14.0
pytest-metadata==3.1.1
pytest-cov==5.0.0

# Reporting
allure-pytest==2.15.3

# Automation
selenium==4.41.0
webdriver-manager==4.0.2

# API
requests==2.32.5

# Config
python-dotenv==1.2.2
PyYAML==6.0.3

# Utilities
colorama==0.4.6
packaging==24.1
tenacity==8.5.0